+++ packages/expo-router/CHANGELOG.md
@@ -85,6 +85,7 @@
 - [ios] add internal animateAspectRatioChange prop ([#41511](https://github.com/expo/expo/pull/41511) by [@Ubax](https://github.com/Ubax))
 - Add unstable_navigationEvents to expose global navigation events ([#41706](https://github.com/expo/expo/pull/41706) by [@Ubax](https://github.com/Ubax))
 - [ios] replace print with jsLog in native components ([#41757](https://github.com/expo/expo/pull/41757) by [@Ubax](https://github.com/Ubax))
+- replace custom event emitter, with expo-modules-core version ([#41807](https://github.com/expo/expo/pull/41807) by [@Ubax](https://github.com/Ubax))
 
 ## 6.0.17 - 2025-12-05